use url params for address
check server conf
