+++ src/components/EnterReadOnlyAddress/EnterReadOnlyAddress.tsx
@@ -5,22 +5,8 @@ import { ValidationMessage, Panel } from './styles'
 import { useStores } from '@/stores/utils'
 import { observer } from 'mobx-react'
 import { useHistory } from 'react-router-dom'
-import { ethers } from 'ethers'
 import { PATHS } from '@/routes'
-
-export const validateAddress = (address: string) => {
-  if (address === '') {
-    return false
-  }
-
-  try {
-    ethers.utils.getAddress(address)
-  } catch (e) {
-    return false
-  }
-
-  return true
-}
+import { validateAddress } from '@/utils/utils'
 
 export const EnterReadOnlyAddress: React.FC = observer((props) => {
   const [address, setAddress] = useState()
@@ -39,7 +25,7 @@ export const EnterReadOnlyAddress: React.FC = observer((props) => {
   const checkBalance = () => {
     if (validateAddress(address!)) {
       appStore.setAddress(address)
-      history.push(PATHS.checkBalance)
+      history.push(PATHS.checkBalance.replace(':address', address))
     } else {
       setIsNotValid(true)
     }
