+++ src/utils/utils.ts
@@ -1,3 +1,5 @@
+import { ethers } from 'ethers'
+
 const currencyFormatter = (currency: string) =>
   new Intl.NumberFormat('en-US', {
     style: 'currency',
@@ -24,4 +26,18 @@ export const prettyEthAddress = (address: string) => {
     address = `${address.substring(0, 6)}...${address.substring(42, 38)}`
   }
   return address
+}
+
+export const validateAddress = (address: string) => {
+  if (address === '') {
+    return false
+  }
+
+  try {
+    ethers.utils.getAddress(address)
+  } catch (e) {
+    return false
+  }
+
+  return true
 }
